Autenticación y Autorización para endpoints de /api/projects

+++ src/controllers/ProjectController.ts
@@ -5,7 +5,13 @@ import colors from "colors"
 export class ProjectController {
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({});
+            const projects = await Project.find({
+                $or: [
+                    {
+                        manager: {$in: req.user.id}
+                    }
+                ]
+            });
             res.json({projects});
         } catch (error) {
             console.log(error);
@@ -22,6 +28,11 @@ export class ProjectController {
                 const error = new Error(`Proyecto con id=${id} no encontrado`)
                 return res.status(404).json({error: error.message})
             }
+            
+            if(project.manager.toString() !== req.user.id.toString()){
+                const error = new Error(`Acción no válida`)
+                return res.status(401).json({error: error.message})
+            }
             res.json({project});
         } catch (error) {
             console.log(error);
@@ -31,6 +42,8 @@ export class ProjectController {
 
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body);
+
+        project.manager = req.user.id;
 
         try {
             // Otra opción: await Project.create(req.body);
@@ -52,6 +65,11 @@ export class ProjectController {
                 return res.status(404).json({error: error.message})
             }
 
+            if(project.manager.toString() !== req.user.id.toString()){
+                const error = new Error(`No tiene permiso para actualizar el proyecto`);
+                return res.status(401).json({error: error.message});
+            }
+
             project.projectName = req.body.projectName;
             project.clientName = req.body.clientName;
             project.description = req.body.description;
@@ -70,8 +88,13 @@ export class ProjectController {
             const project = await Project.findById(id);
 
             if(!project){
-                const error = new Error(`Proyecto con id=${id} no encontrado`)
-                return res.status(404).json({error: error.message})
+                const error = new Error(`Proyecto con id=${id} no encontrado`);
+                return res.status(404).json({error: error.message});
+            }
+
+            if(project.manager.toString() !== req.user.id.toString()){
+                const error = new Error(`No tiene permiso para eliminar el proyecto`);
+                return res.status(401).json({error: error.message});
             }
 
             await project.deleteOne();
